% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{model}
\alias{model}
\title{Bayesian Growth Curve Model for Complete Data, MCAR (Missing Completely at Random,
and MAR (Missing at Random}
\format{
A character string.
}
\usage{
model
}
\description{
A character string containing the JAGS model specification for
complete data (no missing values), MCAR, and MAR.
}
\keyword{datasets}
