% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MNAR.R
\docType{data}
\name{model_MNAR}
\alias{model_MNAR}
\title{Bayesian Growth Curve Model for MNAR (Missing Not At Random)}
\format{
A character string.
}
\usage{
model_MNAR
}
\description{
JAGS model definition for data with MNAR mechanism.
}
\keyword{datasets}
