% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_MNAR_k.R
\docType{data}
\name{model_MNAR_k}
\alias{model_MNAR_k}
\title{Bayesian Growth-Curve Model for MNAR with Auxiliary Variable (k)}
\format{
A character string.
}
\usage{
model_MNAR_k
}
\description{
JAGS model definition for MNAR mechanism with auxiliary variable k.
}
\keyword{datasets}
